+++ src/App.tsx
@@ -1,9 +1,9 @@
+import { BottomLeftPill } from './components/BottomLeftPill';
 import { Canvas } from './components/Canvas/Canvas';
 import { FloatingPlayPause } from './components/FloatingPlayPause';
 import { Header } from './components/Header';
 import { LeftSideBar } from './components/SideBar';
 import { Toolbar } from './components/Toolbar/Toolbar';
-import { ZoomPill } from './components/ZoomPill';
 
 export default function App() {
   return (
@@ -13,7 +13,7 @@ export default function App() {
       <LeftSideBar />
       <Toolbar />
       <FloatingPlayPause />
-      <ZoomPill />
+      <BottomLeftPill />
     </div>
   );
 }
